requirements.txt
asgiref==3.4.1
Django==3.2.18
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.1.1
